pyproject.toml
[build-system]
requires = ["setuptools>=75.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hatchery"
version = "0.1.0"
description = "Docker-based malware sandbox engine — watch it hatch, watch it burn."
readme = "README.md"
requires-python = ">=3.12"
license = "MIT"
authors = [
    {name = "Raphael Main + Agent Mackenzie"},
]

dependencies = [
    "click>=8.1",
    "rich>=13.9",
    "docker>=7.0",
    "pefile>=2024.0.0",
    "pyelftools>=0.32",
    "yara-python>=4.5",
    "flare-capa>=9.0",
    "aiofiles>=24.1",
    "httpx>=0.28",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.25",
]

[project.scripts]
hatchery = "engine.cli:cli"

[tool.setuptools.packages.find]
include = ["engine*"]

[tool.setuptools.package-data]
engine = ["static/rules/**/*.yar", "sandbox/seccomp.json", "sandbox/docker/*"]